+++ CHANGES.rst
@@ -3,6 +3,11 @@ Changelog
 *********
 
 
+development
+===========
+- Add example for annotating phenology data within Grafana
+
+
 2020-12-27 0.1.2
 ================
 - Adjust documentation